+++ R/bed_random.r
@@ -3,10 +3,10 @@
 #' @param genome [tbl_genome()]
 #' @param length length of intervals
 #' @param n number of intervals to generate
-#' @param sort_by sorting variables
 #' @param seed seed RNG for reproducible intervals
+#' @param sorted return sorted output
 #'
-#' @details Sorting can be suppressed with `sort_by = NULL`.
+#' @details Sorting can be suppressed with `sorted = FALSE`.
 #'
 #' @return [tbl_interval()]
 #'
@@ -26,14 +26,13 @@
 #' bed_random(genome, seed = 10104)
 #'
 #' # sorting can be suppressed
-#' bed_random(genome, sort_by = NULL, seed = 10104)
+#' bed_random(genome, sorted = FALSE, seed = 10104)
 #'
 #' # 500 random intervals of length 500
 #' bed_random(genome, length = 500, n = 500, seed = 10104)
 #'
 #' @export
-bed_random <- function(genome, length = 1000, n = 1e6,
-                       sort_by = c("chrom", "start"), seed = 0) {
+bed_random <- function(genome, length = 1000, n = 1e6, seed = 0, sorted = TRUE) {
   if (!is.tbl_genome(genome)) genome <- as.tbl_genome(genome)
 
   if (!all(genome$size > length)) {
@@ -42,9 +41,8 @@ bed_random <- function(genome, length = 1000, n = 1e6,
 
   out <- random_impl(genome, length, n, seed)
 
-  if (!is.null(sort_by) && length(sort_by) > 0) {
-    sort_syms <- rlang::syms(sort_by)
-    out <- arrange(out, !!! sort_syms)
+  if (sorted) {
+    out <- bed_sort(out)
   }
 
   out <- as.tbl_interval(out)